+++ src/ballistica/base/support/app_config.h
@@ -47,6 +47,7 @@ class AppConfig {
     kVRHeadRelativeAudio,
     kMacControllerSubsystem,
     kDevConsoleActiveTab,
+    kInsecureConnections,
     kLast  // Sentinel.
   };